spark-jobs/requirements.txt
pyspark==3.4.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
loguru==0.7.2
pandas==2.0.3
numpy==1.24.4
pyarrow==12.0.1 